reco-system/requirements.txt
streamlit_folium
folium
pandas
pandas-gbq
google-cloud-bigquery
geopy